new autocomplete for new search bar
this gets rid of the names being stored in the JS file. The names are now stored on the chapters page, along with the images and contact and city/state. It is a much smaller file since it doesn't have to import HTML to the chapters file, it only shows or hides items.