MAJOR FIX
I am goated at html

notice how the hrefs are replaced with parent directory shorthand... very demure... very mindful...
basically when i moved the blogs in their own folder all of the references stopped working even though I had changed them to include the > blogs > (specific blog)

but im guessing everything broke because now these inner files don't have access to "fai-main" directly so error 404 happens. you can go to the blog but returning to the blog didn't work and the bg image + arrow in the blog didn't appear

how i fixed this was renaming all instances of "fai-main" to "../" which apprently means to step back a parent, so it doesn't matter what the parent file is named. I don't know why this change is necessary because it worked in vscode but i really hope it fixes it here...